+++ src/inc/uartSendStr.h
@@ -5,7 +5,7 @@
 #ifndef PITOOLS_LIB_UARTSENDSTR_H
 #define PITOOLS_LIB_UARTSENDSTR_H
 
-extern uint8_t uart_idx;
+extern uint16_t uart_idx;
 
 void uartSendStr(const char* str);
